+++ src/test/java/ch/bitagent/bitcoin/lib/ecc/SignatureTest.java
@@ -1,6 +1,6 @@
 package ch.bitagent.bitcoin.lib.ecc;
 
-import ch.bitagent.bitcoin.lib.helper.Helper;
+import ch.bitagent.bitcoin.lib.helper.Bytes;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -18,8 +18,8 @@ class SignatureTest {
         assertTrue(s.eq(sig2.getS()));
 
         for (int i = 0; i < 2; i++) {
-            r = Hex.parse(Helper.randomBytes(Int.parse(2).pow(Int.parse(256)).toBytes(32).length));
-            s = Hex.parse(Helper.randomBytes(Int.parse(2).pow(Int.parse(255)).toBytes(32).length));
+            r = Hex.parse(Bytes.randomBytes(Int.parse(2).pow(Int.parse(256)).toBytes(32).length));
+            s = Hex.parse(Bytes.randomBytes(Int.parse(2).pow(Int.parse(255)).toBytes(32).length));
             sig1 = new Signature(r, s);
             der = sig1.der();
             sig2 = Signature.parse(der);