+++ release-notes.md
@@ -1,16 +1,19 @@
 # release notes
 
+## snapshot
+- mnemonic sentence (bip-0039)
+
 ## 0.1.0
 - wallet
-- extended key
+- extended key (bip-0084)
 - refactoring address (breaking change)
 - refactoring hash (breaking change)
 
 ## 0.0.5
-- schnorr
+- schnorr (bip-0340)
 
 ## 0.0.4
-- bech32
+- bech32 (bip-0173, bip-0350)
 - address util 
 
 ## 0.0.3